rename the wrapper-based withErrorModal
