+++ ui/src/app/components/with-error-modal-wrapper.tsx
@@ -15,7 +15,7 @@ export interface WithErrorModalProps {
   showErrorModal: (title: string, body: string) => void;
 }
 
-export const withErrorModal = () => {
+export const withErrorModalWrapper = () => {
   return (WrappedComponent) => {
     return class ErrorModalWrapper extends React.Component<any, State> {
       constructor(props) {
@@ -76,7 +76,7 @@ export const withProfileErrorWrapper = (WrappedComponent) => {
     </React.Fragment>
   );
 
-  const ProfileErrorWrapper = ({ showErrorModal, ...props }) => {
+  return ({ showErrorModal, ...props }) => {
     return (
       <WrappedComponent
         showProfileErrorModal={(message) =>
@@ -86,11 +86,9 @@ export const withProfileErrorWrapper = (WrappedComponent) => {
       />
     );
   };
-
-  return ProfileErrorWrapper;
 };
 
 export const withProfileErrorModal = fp.flow(
   withProfileErrorWrapper,
-  withErrorModal()
+  withErrorModalWrapper()
 );
