+++ ui/src/app/pages/analysis/notebook-resource-card.tsx
@@ -19,9 +19,9 @@ import {
   WithConfirmDeleteModalProps,
 } from 'app/components/with-confirm-delete-modal';
 import {
-  withErrorModal,
   WithErrorModalProps,
-} from 'app/components/with-error-modal';
+  withErrorModalWrapper,
+} from 'app/components/with-error-modal-wrapper';
 import {
   withSpinnerOverlay,
   WithSpinnerOverlayProps,
@@ -49,7 +49,7 @@ interface State {
 }
 
 export const NotebookResourceCard = fp.flow(
-  withErrorModal(),
+  withErrorModalWrapper(),
   withConfirmDeleteModal(),
   withSpinnerOverlay()
 )(
