+++ ui/src/app/pages/data/cohort/cohort-resource-card.tsx
@@ -19,9 +19,9 @@ import {
   WithConfirmDeleteModalProps,
 } from 'app/components/with-confirm-delete-modal';
 import {
-  withErrorModal,
   WithErrorModalProps,
-} from 'app/components/with-error-modal';
+  withErrorModalWrapper,
+} from 'app/components/with-error-modal-wrapper';
 import {
   withSpinnerOverlay,
   WithSpinnerOverlayProps,
@@ -54,7 +54,7 @@ interface State {
 }
 
 export const CohortResourceCard = fp.flow(
-  withErrorModal(),
+  withErrorModalWrapper(),
   withConfirmDeleteModal(),
   withSpinnerOverlay(),
   withNavigation
