+++ ui/src/app/pages/data/data-set/dataset-page.tsx
@@ -39,9 +39,9 @@ import { CheckBox } from 'app/components/inputs';
 import { TooltipTrigger } from 'app/components/popups';
 import { Spinner, SpinnerOverlay } from 'app/components/spinners';
 import {
-  withErrorModal,
   WithErrorModalProps,
-} from 'app/components/with-error-modal';
+  withErrorModalWrapper,
+} from 'app/components/with-error-modal-wrapper';
 import { WithSpinnerOverlayProps } from 'app/components/with-spinner-overlay';
 import { CircleWithText } from 'app/icons/circleWithText';
 import { ExportDatasetModal } from 'app/pages/data/data-set/export-dataset-modal';
@@ -786,7 +786,7 @@ export const DatasetPage = fp.flow(
   withUserProfile(),
   withCurrentWorkspace(),
   withCdrVersions(),
-  withErrorModal(),
+  withErrorModalWrapper(),
   withRouter
 )(
   ({
