+++ ui/src/app/pages/data/data-set/dataset-resource-card.tsx
@@ -19,9 +19,9 @@ import {
   WithConfirmDeleteModalProps,
 } from 'app/components/with-confirm-delete-modal';
 import {
-  withErrorModal,
   WithErrorModalProps,
-} from 'app/components/with-error-modal';
+  withErrorModalWrapper,
+} from 'app/components/with-error-modal-wrapper';
 import {
   withSpinnerOverlay,
   WithSpinnerOverlayProps,
@@ -56,7 +56,7 @@ interface State {
 }
 
 export const DatasetResourceCard = fp.flow(
-  withErrorModal(),
+  withErrorModalWrapper(),
   withConfirmDeleteModal(),
   withSpinnerOverlay(),
   withNavigation
