+++ ui/src/app/pages/workspace/workspace-about.tsx
@@ -208,6 +208,7 @@ export const WorkspaceAbout = fp.flow(
 
     loadUserRoles(workspace: WorkspaceData) {
       this.setState({ workspaceUserRoles: [] });
+
       workspacesApi()
         .getFirecloudWorkspaceUserRoles(workspace.namespace, workspace.id)
         .then((resp) => {
